restrict usage to bot commands channel and fix header sending on all argument

+++ commands/thaler.js
@@ -2,21 +2,27 @@ const { data } = require("../guides/thalerGuide.js");
 
 exports.run = async (client, message, args, level) => {
 
-  await message.channel.send({
-    files: [
-    {
-    attachment: 'media/img/guides/break.png',
-    name: 'break.png'
-    },
-    {
-    attachment: 'media/img/guides/thalerheader.png',
-    name: 'AFK Thaler header.png'
-    },
-    {
-    attachment: 'media/img/guides/break.png',
-    name: 'break.png'
-    }]
-  });
+  if (message.channel.id !== '382701090430386180' && level < 2) return;
+
+    await message.channel.send({
+      files: [{
+      attachment: 'media/img/guides/break.png',
+      name: 'break.png'
+      }]
+    });
+    await message.channel.send({
+      files: [{
+      attachment: 'media/img/guides/thalerheader.png',
+      name: 'AFK Thaler header.png'
+      }]
+    });
+    await message.channel.send({
+      files: [{
+      attachment: 'media/img/guides/break.png',
+      name: 'break.png'
+      }]
+    });
+      
 	await message.channel.send('', {embed: data[0]});
 	await message.channel.send('', {embed: data[1]});
 	message.delete();
